Move RPC interfaces to their own sub-package name
